add geometry validator, fix ring linearization, merge element handlers
IIGeometryValidator: Remove a lot of checks that are not necessary
because they are handled by the JTS validation. Thereby also remove a
System.out.println(..) statement that could produce unnecessary log
entries. The geometry validator now uses the custom curve linearizer,
with linearization criterion as defined for the GmlGeoX instance.

Ring linearization: The IIGeometryFactory now creates IIRing objects,
which implement a buildJTSGeometry() method that uses the custom curve
linearizer. The DefaultRing implementation from deegree did not use that
linearization.

Merging of element handlers: The
SecondaryGeometryElementValidationHandler has been merged with the
GeometryElementHandler. The latter now also covers the checks of the
former (which has been removed). The primary reason was to re-use
computed JTS geometries for the different validation checks as much as
possible. GeometricValidationImpl.java has been removed as well, since
it was not used within GmlGeoX.



Signed-off-by: Johannes Echterhoff <echterhoff@interactive-instruments.de>